implement activity routes

+++ models/instructor/room/RoomRoutes.php
@@ -1,6 +1,8 @@
 <?php
 require_once('./models/instructor/room/RoomService.php');
 require_once('./models/exception/NotFoundException.php');
+require_once('./models/instructor/activity/ActivityRoutes.php');
+
 
 
 class RoomRoutes
@@ -14,10 +16,12 @@ class RoomRoutes
         $this->request_data = $request_data;
         $this->middleware = $middleware;
 
+
+        $trimmed_url = str_replace('room/', '', $this->request_data->get_request_url());
+        $this->request_data->set_url($trimmed_url);
+
         $this->url = explode('/', $this->request_data->get_request_url());
 
-        $trimmed_url = str_replace('instructor/', '', $this->request_data->get_request_url());
-        $this->request_data->set_url($trimmed_url);
 
         $this->method = $this->request_data->get_request_method();
         $this->room_service = new RoomService();
@@ -25,7 +29,6 @@ class RoomRoutes
 
     public function handle_url()
     {
-
         $request_body = json_decode(file_get_contents('php://input'));
 
         $params = $this->request_data->get_request_params();
@@ -33,49 +36,56 @@ class RoomRoutes
         $count = count($url);
         $current_route = $url[0];
         $next_route = null;
-
-        if($count > 1){
+        
+        if ($count > 1) {
             $next_route = $url[1];
         }
 
-        switch ($this->method) {
-            case 'POST':
-                if ($current_route == 'room' && $count == 0) {
-                    echo json_encode($this->room_service->add_room($request_body, $this->middleware->get_owner_email()));
-                }
-                break;
-            case 'GET':
-                if ($params) {
-                    echo json_encode($this->room_service->search_rooms($params['search']));
-                    echo $params['search'];
-                }
+        if($current_route == 'activity' || $current_route == 'activities'){
 
-                if ($current_route == 'room' && $count == 2) {
-                    
+            $activity_routes = new ActivityRoutes($this->request_data, $this->middleware);
+            $activity_routes->handle_url();
+
+        }else{
+
+            switch ($this->method) {
+                case 'POST':
+                    if ($current_route == 'room' && $count == 0) {
+                        echo json_encode($this->room_service->add_room($request_body, $this->middleware->get_owner_email()));
+                    }
+                    break;
+                case 'GET':
+                    if ($params) {
+                        echo json_encode($this->room_service->search_rooms($params['search']));
+                    }
+
+                    if ($current_route == 'room' && $count == 2) {
                         if (intval($next_route) > 0) {
                             echo json_encode($this->room_service->get_room($next_route));
                         } else {
                             throw new NotFoundException();
                         }
-                    
-                    
-                }else if($current_route == 'rooms'){
-                    echo json_encode($this->room_service->get_rooms($this->middleware->get_owner_id()));
-                }
-                break;
-            case 'PUT':
-                if ($current_route == 'room') {
-                    echo json_encode($this->room_service->update_room($request_body, $next_route));
-                }
-                break;
-            case 'DELETE':
-                if ($current_route == 'room') {
-                    echo json_encode($this->room_service->delete_room($next_route));
-                }
-                break;
-            default:
-                throw new NotFoundException();
-                break;
+                    } else if ($current_route == 'rooms') {
+                        echo json_encode($this->room_service->get_rooms($this->middleware->get_owner_id()));
+                    }
+                    break;
+                case 'PUT':
+                    if ($current_route == 'room') {
+                        echo json_encode($this->room_service->update_room($request_body));
+                    }
+                    break;
+                case 'DELETE':
+                    if ($current_route == 'room') {
+                        echo json_encode($this->room_service->delete_room($next_route));
+                    }
+                    break;
+                default:
+                    throw new NotFoundException();
+                    break;
+            }
+            
         }
+
+        
     }
 }
